+++ waves/wcore/templatetags/waves_tags.py
@@ -10,8 +10,10 @@ register = template.Library()
 @register.inclusion_tag('waves/services/_online_execution.html', takes_context=True)
 def online_exec_button(context, service, label=None):
     """ for service, setup if current usr can submit jobs """
-    return {'available_for_submission': service.available_for_user(context['user']) and context.get('preview') is None,
-            'label': label, 'service': service}
+    if service.default_submission:
+        return {'available_for_submission': service.default_submission.available_for_user(context['user']) and context.get('preview') is None,
+                'label': label, 'service': service}
+    return {}
 
 
 @register.simple_tag